Multiple Screen changes
Added some screen classes to handle the mulitple screens we will have
to use to play the game.
Starting with just a splash screen that fades in and out, and then the
Game screen, which just plays the game.

We can add more as needed now that the framework is there.